pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "streamlit",
    "python-dotenv",
    "langchain",
    "langchain-core",
    "langchain-community",
    "langchain-openai",
    "langchain-text-splitters",
    "langgraph",
    "langgraph-prebuilt",
    "arxiv",
    "chromadb",
    "pypdf",
    "youtube-transcript-api>=1.2.4",
    "langchain-chroma>=1.1.0",
    "wikipedia",
    "langgraph-checkpoint-postgres",
    "psycopg[binary]",
    "boto3",
]
